util: fix creating nameless fibers
The problem is the fact that a reloadable fiber starts before the name
is set. So at the first iteration of executing its name is 'lua', which
is the default value for `fiber.create` or `fiber.new`.

In order to save logs and states we need fiber name to be set properly.
Let's use `fiber.new` and not `fiber.create` in order to initialize the
fiber. It won't start before context switching is performed. This way
we will have time to set the fiber's name.

I also think that we should yield in the end in order not to change
the API of utils: `fiber_create` is creating and starting a fiber.
`fiber.yield` will give the fiber some time to start its execution.

Needed for #107