log_fiber: introduce the module for saving logs
This patch introduces the new module, the purpose of which is to save the
last log messages and the states of the vshard's background fibers.

The main motivation is convenience of writing luatest tests for
background fibers using this module and not manual grepping of logs as
it is done in diff tests.

Part of #107

@TarantoolBot document
Title: new option for vshard's cfg

log_vshard_background

Since version 0.1.21. Specifies the number of messages saved in the
logger of the background fibers, which can be acessed via
`vshard.router/storage.info()`.

This option can be of boolean type. In that case, if set to true,
default value of saved messages (5) will be used. You can also specify
this number manually.

In case of reconfiguration of the parameter from 'true' or any
non-negative number to `false` or `nil` all saved messages and states
of the fibers will be dropped.

Type: boolean or number
Default: no
Dynamic: true