storage: add saving of logs for background fibers
This patch introduce the ability for the user to enable saving logs
and states from the background fibers. This information can be acessed
via 'vshard.storage.info()'.

The problem is the fact that background fibers are difficult to test
using luatest, as there's no default way to access logs during test
execution. This patch is intended to solve that issue.

Closes of #107