ci: fix using cached Tarantool master branch
When the cache is found on runner, running tests fails with the
following error: "libicui18n.so.66: cannot open shared object file".

The problem here, that on cached workflow the libraries, used
during build are not installed and consequently cannot be linked with.

Let's install all needed dependencies even when cache is found.

NO_DOC=ci
NO_TEST=ci