perf: log the process of performance test
This commit adds several log messages, so that it's easier to see, what
is going on with the test. It may take huge amount of time, when serious
load is generated.

While we're here, let's also fix the error printing. If not vshard call
is used, then the return value may be box.NULL. In such case the test
spams "cdata<void *>: NULL" in log. Let's explicitly check, that it's
not 'nil' and pring it only after that.

NO_DOC=perf
NO_TEST=perf