error: introduce the function for checking whether object is error
In the following commit the function, which checks, whether the
table is error, will be needed. This cannot be done solely with
`box.error.is` API, since vshard unpacks all errors to tables by
default.

We consider a table as error, if it either tarantool's error or
table with type and code.

Needed for #578
Needed for #583
Needed for #588

NO_DOC=internal