replicaset: do not spam with errors in logs on every failed call
Users are worried about the space on disk, which logs might use in case
of instance failure and we should relax the verbosity of logs for that.

`replica_call` in the replicaset module logged the error every time a
request fail, which produced a lot of logs, when instance is down. From
now on the errors with same type and code won't be logged.

The limiter must be flushed periodically, since otherwise, if e.g. 100
errors happened in 10 seconds and then stopped, the 'Suppressed' message
won't be printed until the next error. And this flush cannot happen
on success path of the `replica_call`, since that'll affect performance.

Part of #588

NO_DOC=bugfix