storage: deduplicate error logs in services
Recovery, gc and rebalancer service used to emit an error on every
iteration while an instance was unreachable, flooding logs with the same
message. From now on this won't be the case, same errors are not printed
anymore. Moreover, some errors are not printed at all (e.g.
STORAGE_IS_DISABLED or VHANDSHAKE_NOT_COMPLETE).

The new test file is introduced, since the tests are slow and I'd like
for them to run in parallel with others.

Part of #578
Part of #588

NO_DOC=bugfix